test: remove obsolete migration and internal layout assertions
